# swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "To-do app API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/todo-list/",
      "description": "Основной сервер"
    }
  ],
  "paths": {
    "/tasks": {
      "post": {
        "summary": "Создать новое задание",
        "description": "Этот запрос нужен для создания новой задачи:\n  * Может вернуть ошибку, если данные не валидны\n  * Может вернуть ошибку, если произошла ошибка на стороне сервера",
        "tags": [
          "tasks"
        ],
        "operationId": "CreateTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешный ответ сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskId"
                }
              }
            }
          },
          "400": {
            "description": "Возвращает если данные введенные пользователем не валидные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Все нестандартное",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Получить список задач",
        "description": "Этот запрос нужен для получения списка задач , присуствует фильтрация по статусу (по умолчанию: active)\n  * Может вернуть ошибку, если данные не валидны\n  * Может вернуть ошибку, если произошла ошибка на стороне сервера",
        "tags": [
          "tasks"
        ],
        "operationId": "GetTasksByStatus",
        "parameters": [
          {
            "name": "status",
            "description": "Статус задачи, может принимать только active и done",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "done"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ сервера со списком отфильтрованных задач",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tasks"
                }
              }
            }
          },
          "400": {
            "description": "Возвращает если данные введенные пользователем не валидны",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Возвращает код, если список задач пустой",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tasks"
                }
              }
            }
          },
          "default": {
            "description": "Все нестандартное",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Получить информацию о задаче",
        "description": "Этот запрос нужен для получения информации о конкретной задаче по её идентификатору.\n  * Может вернуть ошибку, если идентификатор невалиден.\n  * Может вернуть ошибку, если задача не найдена.\n  * Может вернуть ошибку, если произошла ошибка",
        "operationId": "GetTask",
        "tags": [
          "tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Идентификатор задачи",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ сервера с нужной задачей",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Возвращает если данные введенные пользователем не валидны",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Возвращает код, если задача не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Все нестандартное",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Обновить задачу",
        "description": "Этот запрос нужен для изменения данных задачи\n  * Может вернуть ошибку, если данные не валидны\n  * Может вернуть ошибку, если произошла ошибка на стороне сервера\n  * Может вернуть ошибку, если задача не найдена",
        "operationId": "UpdateTask",
        "tags": [
          "tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Идентификатор задачи которую нужно обновить",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Успешный ответ сервера"
          },
          "400": {
            "description": "Возвращает если данные введенные пользователем не валидны",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Возвращает код, если задача не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Все нестандартное",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Удалить задачу",
        "tags": [
          "tasks"
        ],
        "description": "Этот метод удаляет задачу из сервера\n  * Может вернуть ошибку, если данные не валидны\n  * Может вернуть ошибку, если произошла ошибка на стороне сервера\n  * Может вернуть ошибку, если задача не нашлась на стороне сервера",
        "operationId": "DeleteTask",
        "parameters": [
          {
            "name": "id",
            "description": "Идентификатор задачи которую нужно удалить",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешный ответ сервера"
          },
          "400": {
            "description": "Возвращает если данные введенные пользователем не валидны",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Возвращает код, если задача не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Все нестандартное",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}/done": {
      "put": {
        "summary": "Пометить задание выполненным",
        "description": "Этот запрос нужен для обновления статуса задачи на \"done\"\n  * Может вернуть ошибку, если данные не валидны\n  * Может вернуть ошибку, если произошла ошибка на стороне сервера\n  * Может вернуть ошибку, если статус задачи уже изменен",
        "operationId": "FinishTask",
        "tags": [
          "tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Идентификатор задачи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Успешный ответ сервера"
          },
          "default": {
            "description": "Все нестандартное",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "required": [
          "title",
          "activeAt"
        ],
        "properties": {
          "id": {
            "description": "Идентификатор задачи",
            "type": "string",
            "example": "67d51f98a11d11a3cce57c7c"
          },
          "title": {
            "description": "Описание задачи",
            "type": "string",
            "example": "Купить новую машину"
          },
          "activeAt": {
            "description": "Дедлайн задачи",
            "type": "string",
            "example": "2023-08-24T00:00:00.000Z",
            "format": "2023-08-24T00:00:00.000Z"
          },
          "status": {
            "description": "Статус задачи",
            "type": "string",
            "example": "active"
          }
        }
      },
      "Tasks": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Task"
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "description": "Описание ошибки",
            "type": "string",
            "example": "id field is empty"
          }
        }
      },
      "TaskId": {
        "type": "object",
        "required": [
          "inserted_id"
        ],
        "properties": {
          "inserted_id": {
            "description": "Идентификатор задачи",
            "type": "string",
            "example": "67d51f98a11d11a3cce57c7c"
          }
        }
      }
    }
  }
}